Separate out into more controllers.
This allows us to show off event triggers and a way of getting
information from one controller to another.
