+++ samples/mapping/app/styles/map.tss
@@ -1,0 +1,12 @@
+"#map": {
+    top: '50dp',
+    animate: true,
+    regionFit: true,
+    userLocation: false,
+    region: {
+        latitude: Alloy.Globals.LATITUDE_BASE,
+        longitude: Alloy.Globals.LONGITUDE_BASE,
+        latitudeDelta: 0.1, 
+        longitudeDelta: 0.1
+    }
+}
