+++ samples/mapping/app/views/map.xml
@@ -1,0 +1,5 @@
+<Alloy>
+    <View ns="Ti.Map" id="map" mapType="Ti.Map.STANDARD_TYPE">
+        <Require src="annotation" title="Appcelerator" />
+    </View>
+</Alloy>
